x-pack/filebeat/input/azureblobstorage: report Running after a recovered empty poll
Follow-up to review feedback on the blob-listing resilience work.

When polling, a transient listing failure marks the input Degraded and
returns without error so it can retry on the next poll. Previously the
input only returned to Running once a job published an event, so a poll
that recovered but scheduled no jobs (no new blobs, or everything
filtered out or already past the checkpoint) could leave the input stuck
in Degraded even though listing had recovered. Report Running after a
successful listing pass that schedules no jobs so recovery is signalled
on an empty poll as well.

Change retryConfig.MaxRetries from int to int32 to match the Azure SDK's
policy.RetryOptions and the download stream reader, removing the
redundant conversions at the call sites.

Migrate the package's tests off the deprecated logp.TestingSetup to
localized logptest loggers, and add a test covering the
Degraded -> successful empty poll -> Running recovery.

Document the stack versions the retry attribute is available in.

Updates #44629

+++ x-pack/filebeat/input/azureblobstorage/client.go
@@ -40,7 +40,7 @@ func fetchServiceClientAndCreds(cfg config, retryCfg retryConfig, url string, lo
 // defaultConfig, so a zero value here only occurs when explicitly configured.
 func azureRetryOptions(rc retryConfig) policy.RetryOptions {
 	return policy.RetryOptions{
-		MaxRetries:    int32(rc.MaxRetries),
+		MaxRetries:    rc.MaxRetries,
 		RetryDelay:    rc.InitialRetryDelay,
 		MaxRetryDelay: rc.MaxRetryDelay,
 	}
